+++ src/main/java/org/cascadebot/cascadebot/messaging/MessagingTimed.java
@@ -1,0 +1,55 @@
+/*
+ * Copyright (c) 2019 CascadeBot. All rights reserved.
+ * Licensed under the MIT license.
+ */
+
+package org.cascadebot.cascadebot.messaging;
+
+import net.dv8tion.jda.core.entities.Message;
+import net.dv8tion.jda.core.entities.MessageEmbed;
+import net.dv8tion.jda.core.utils.Checks;
+import org.cascadebot.cascadebot.commandmeta.CommandContext;
+
+public class MessagingTimed {
+
+    CommandContext context;
+    public MessagingTimed(CommandContext context) {
+        this.context = context;
+    }
+
+    /**
+     * Sends a message that auto deletes it's self after the specified delay (in mills).
+     *
+     * @param message The string message to send which cannot be blank.
+     * @param delay   The amount of time to wait before it deletes it's self.
+     * @throws IllegalArgumentException if message is blank.
+     */
+    public void sendAutoDeleteMessage(String message, long delay) {
+        Checks.notBlank(message, "message");
+        Messaging.sendAutoDeleteMessage(context.getChannel(), message, delay);
+    }
+
+    /**
+     * Sends a message that auto deletes it's self after the specified delay (in mills).
+     *
+     * @param embed The non-null {@link MessageEmbed} object to send.
+     * @param delay The amount of time to wait before it deletes it's self.
+     */
+    public void sendAutoDeleteMessage(MessageEmbed embed, long delay) {
+        Checks.notNull(embed, "embed");
+        Messaging.sendAutoDeleteMessage(context.getChannel(), embed, delay);
+    }
+
+    /**
+     * Sends a message that auto deletes it's self after the specified delay (in mills).
+     *
+     * @param message The non-null {@link Message} object to send.
+     * @param delay   The amount of time to wait before it deletes it's self.
+     * @throws IllegalArgumentException if message is null.
+     */
+    public void sendAutoDeleteMessage(Message message, long delay) {
+        Checks.notNull(message, "message");
+        Messaging.sendAutoDeleteMessage(context.getChannel(), message, delay);
+    }
+
+}
